A web page can now link to resources like css and js, provided by a module
a new action, jelix~www:getfile allow to retrieve a file stored in a www directory inside a module. It takes two parameters: targetmodule and file.
New plugins for jtpl allow to generate links easily: meta_htmlmodule to add links into the header of an html page. And resurl to retrieve links for other html tag such as img.
The module junittests has been adapted as an example.